+++ src/net/command/item/CommandPotion.java
@@ -13,7 +13,7 @@ public class CommandPotion extends Command {
 		Connection connection = player.getConnection();
 		int id = connection.readInt();
 		if(PotionManager.exists(id)) {
-			connection.writeByte(PacketID.POTION);
+			connection.writeShort(PacketID.POTION);
 			connection.writePotion(PotionManager.getPotion(id));
 			connection.send();
 		}
