+++ src/net/command/item/CommandRequestItem.java
@@ -16,25 +16,25 @@ public class CommandRequestItem extends Command {
 		Connection connection = player.getConnection();
 		int id = connection.readInt();
 		if(StuffManager.exists(id)) {
-			connection.writeByte(PacketID.STUFF);
+			connection.writeShort(PacketID.STUFF);
 			connection.writeStuff(StuffManager.getStuff(id));
 			connection.send();
 			player.addItemSentToClient(id);
 		}
 		else if(WeaponManager.exists(id)) {
-			connection.writeByte(PacketID.WEAPON);
+			connection.writeShort(PacketID.WEAPON);
 			connection.writeWeapon(WeaponManager.getWeapon(id));
 			connection.send();
 			player.addItemSentToClient(id);
 		}
 		else if(GemManager.exists(id)) {
-			connection.writeByte(PacketID.GEM);
+			connection.writeShort(PacketID.GEM);
 			connection.writeGem(GemManager.getGem(id));
 			connection.send();
 			player.addItemSentToClient(id);
 		}
 		else if(PotionManager.exists(id)) {
-			connection.writeByte(PacketID.POTION);
+			connection.writeShort(PacketID.POTION);
 			connection.writePotion(PotionManager.getPotion(id));
 			connection.send();
 			player.addItemSentToClient(id);
