+++ src/net/command/item/CommandStuff.java
@@ -13,7 +13,7 @@ public class CommandStuff extends Command {
 		Connection connection = player.getConnection();
 		int id = connection.readInt();
 		if(StuffManager.exists(id)) {
-			connection.writeByte(PacketID.STUFF);
+			connection.writeShort(PacketID.STUFF);
 			connection.writeStuff(StuffManager.getStuff(id));
 			connection.send();
 		}
